fix: Settings Modal + dynamic version from package.json
- Settings is now displayed in a modal (not full-screen)
- Version number is now dynamically read from package.json instead of hard-coded
- Added SettingsModal component with proper header and close button
